requirements.txt
PyQRCode (1.2.1)
Pillow (8.1.2)
pypng (0.0.20)